Fix markers, actors and refactor time
